requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
accelerate==0.15.0
    # via diffusers
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
    #   watchfiles
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   uvicorn
diffusers[torch]==0.11.1
    # via -r requirements.in
dnspython==2.3.0
    # via email-validator
email-validator==1.3.1
    # via fastapi
fastapi[all]==0.89.1
    # via -r requirements.in
filelock==3.9.0
    # via
    #   diffusers
    #   huggingface-hub
    #   transformers
ftfy==6.1.1
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.16.3
    # via httpx
httptools==0.5.0
    # via uvicorn
httpx==0.23.3
    # via fastapi
huggingface-hub==0.11.1
    # via
    #   diffusers
    #   transformers
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   requests
    #   rfc3986
importlib-metadata==6.0.0
    # via diffusers
itsdangerous==2.1.2
    # via fastapi
jinja2==3.1.2
    # via fastapi
markupsafe==2.1.2
    # via jinja2
mypy-extensions==0.4.3
    # via typing-inspect
numpy==1.24.1
    # via
    #   accelerate
    #   diffusers
    #   scipy
    #   transformers
    #   xformers
orjson==3.8.5
    # via fastapi
packaging==23.0
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
pillow==9.4.0
    # via diffusers
psutil==5.9.4
    # via accelerate
pydantic==1.10.4
    # via fastapi
pyre-extensions==0.0.23
    # via xformers
python-dotenv==0.21.1
    # via uvicorn
python-multipart==0.0.5
    # via fastapi
pyyaml==6.0
    # via
    #   accelerate
    #   fastapi
    #   huggingface-hub
    #   transformers
    #   uvicorn
regex==2022.10.31
    # via
    #   diffusers
    #   transformers
requests==2.28.2
    # via
    #   diffusers
    #   huggingface-hub
    #   transformers
rfc3986[idna2008]==1.5.0
    # via httpx
scipy==1.10.0
    # via -r requirements.in
six==1.16.0
    # via python-multipart
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.22.0
    # via fastapi
tokenizers==0.13.2
    # via transformers
torch==1.13.1
    # via
    #   accelerate
    #   diffusers
    #   xformers
tqdm==4.64.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.25.1
    # via -r requirements.in
typing-extensions==4.4.0
    # via
    #   huggingface-hub
    #   pydantic
    #   pyre-extensions
    #   torch
    #   typing-inspect
typing-inspect==0.8.0
    # via pyre-extensions
ujson==5.7.0
    # via fastapi
urllib3==1.26.14
    # via requests
uvicorn[standard]==0.20.0
    # via fastapi
watchfiles==0.18.1
    # via uvicorn
wcwidth==0.2.6
    # via ftfy
websockets==10.4
    # via uvicorn
xformers==0.0.16
    # via -r requirements.in
zipp==3.11.0
    # via importlib-metadata
